requirements.txt
aiohttp==3.7.4
alpha-vantage==2.3.1
arabic-reshaper==2.1.1
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
Brotli==1.0.9
category-encoders==2.2.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
dash==1.19.0
dash-bootstrap-components==0.11.3
dash-core-components==1.15.0
dash-extensions==0.0.45
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Caching==1.10.0
Flask-Compress==1.9.0
Flask-SQLAlchemy==2.4.4
fpdf==1.7.2
future==0.18.2
html5lib==1.1
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
kaleido==0.2.1
MarkupSafe==1.1.1
more-itertools==8.7.0
multidict==5.1.0
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.2
parso==0.8.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.1.2
plotly==4.14.3
prompt-toolkit==3.0.16
psutil==5.8.0
Pygments==2.8.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.64
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
SQLAlchemy==1.3.23
statsmodels==0.12.2
threadpoolctl==2.1.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
xgboost==1.3.3
xhtml2pdf==0.2.5
xlrd==2.0.0
yarl==1.6.3
